refactor(diffshub): split diffs hub utilities
Replace the diffsHubUtils aggregate with one named utility file per exported helper.

Update callers and focused tests to import the direct helper modules.
