+++ apps/diffshub/components/DiffUrlForm.tsx
@@ -14,7 +14,7 @@ import {
 import { createPortal } from 'react-dom';
 
 import { Button } from '@/components/Button';
-import { getPatchViewerHref } from '@/lib/diffsHubUtils';
+import { getPatchViewerHref } from '@/lib/getPatchViewerHref';
 import { cn } from '@/lib/utils';
 
 interface DiffUrlFormProps {
