+++ apps/diffshub/lib/getGitHubPath.ts
@@ -1,0 +1,10 @@
+import { getGitHubPathFromURL } from './getGitHubPathFromURL';
+
+export function getGitHubPath(input: string): string | undefined {
+  try {
+    const parsedURL = new URL(input);
+    return getGitHubPathFromURL(parsedURL);
+  } catch {
+    return undefined;
+  }
+}
