+++ apps/diffshub/lib/isDiffItem.ts
@@ -1,0 +1,9 @@
+import type { CodeViewDiffItem, CodeViewItem } from '@pierre/diffs';
+
+import type { CommentMetadata } from './types';
+
+export function isDiffItem(
+  item: CodeViewItem<CommentMetadata>
+): item is CodeViewDiffItem<CommentMetadata> {
+  return item.type === 'diff';
+}
