+++ apps/diffshub/lib/isDraftMetadata.ts
@@ -1,0 +1,7 @@
+import type { CommentMetadata, DraftCommentMetadata } from './types';
+
+export function isDraftMetadata(
+  metadata: CommentMetadata
+): metadata is DraftCommentMetadata {
+  return metadata.kind === 'draft';
+}
