+++ apps/diffshub/lib/isSavedAnnotation.ts
@@ -1,0 +1,9 @@
+import type { DiffLineAnnotation } from '@pierre/diffs';
+
+import type { CommentMetadata, SavedCommentMetadata } from './types';
+
+export function isSavedAnnotation(
+  annotation: DiffLineAnnotation<CommentMetadata>
+): annotation is DiffLineAnnotation<SavedCommentMetadata> {
+  return annotation.metadata.kind === 'saved';
+}
