+++ apps/diffshub/lib/removeSavedCommentSidebarEntry.ts
@@ -1,0 +1,44 @@
+import type {
+  DiffsHubDeletedCommentEvent,
+  DiffsHubSavedCommentItem,
+} from './types';
+
+export function removeSavedCommentSidebarEntry(
+  sections: readonly DiffsHubSavedCommentItem[],
+  entry: DiffsHubDeletedCommentEvent
+): DiffsHubSavedCommentItem[] {
+  let sectionIndex = -1;
+  for (let index = 0; index < sections.length; index++) {
+    if (sections[index]?.itemId === entry.itemId) {
+      sectionIndex = index;
+      break;
+    }
+  }
+
+  if (sectionIndex === -1) {
+    return sections.slice();
+  }
+
+  const section = sections[sectionIndex];
+  if (section == null) {
+    return sections.slice();
+  }
+
+  const nextComments = section.comments.filter(
+    (comment) => comment.key !== entry.key
+  );
+  if (nextComments.length === section.comments.length) {
+    return sections.slice();
+  }
+
+  if (nextComments.length === 0) {
+    return sections.filter((_, index) => index !== sectionIndex);
+  }
+
+  const nextSections = [...sections];
+  nextSections[sectionIndex] = {
+    ...section,
+    comments: nextComments,
+  };
+  return nextSections;
+}
